+++ ICD.Connect.Panels.CrestronPro/Controls/TouchScreens/AbstractThreeSeriesTouchScreenControl.cs
@@ -1,5 +1,4 @@
 ﻿using ICD.Connect.Devices;
-using ICD.Connect.Panels.Controls;
 using ICD.Connect.Panels.Controls.Panels;
 using ICD.Connect.Panels.Crestron.Controls.TouchScreens;