+++ ICD.Connect.Panels.CrestronPro/TriListAdapters/FtTs600/FtTs600Adapter.cs
@@ -1,8 +1,7 @@
 ﻿using ICD.Connect.Conferencing.Controls.Dialing;
-using ICD.Connect.Devices.Controls.Power;
+using ICD.Connect.Panels.Controls.Backlight;
 #if SIMPLSHARP
 using Crestron.SimplSharpPro;
-using ICD.Connect.Devices.Controls;
 using ICD.Connect.Panels.CrestronPro.TriListAdapters.Controls.Backlight;
 using ICD.Connect.Panels.CrestronPro.TriListAdapters.Controls.Voip;
 #endif
@@ -41,7 +40,7 @@ namespace ICD.Connect.Panels.CrestronPro.TriListAdapters.FtTs600
 		/// </summary>
 		/// <param name="id"></param>
 		/// <returns></returns>
-		protected override IPowerDeviceControl InstantiateBacklightControl(int id)
+		protected override IBacklightDeviceControl InstantiateBacklightControl(int id)
 		{
 			return new TswFt5ButtonBacklightControl(this, id);
 		}