+++ ICD.Connect.Panels/Controls/Backlight/BacklightDeviceControlConsole.cs
@@ -1,0 +1,50 @@
+﻿using System;
+using System.Collections.Generic;
+using ICD.Connect.API.Commands;
+using ICD.Connect.API.Nodes;
+
+namespace ICD.Connect.Panels.Controls.Backlight
+{
+	public static class BacklightDeviceControlConsole
+	{
+		/// <summary>
+		/// Gets the child console nodes.
+		/// </summary>
+		/// <param name="instance"></param>
+		/// <returns></returns>
+		public static IEnumerable<IConsoleNodeBase> GetConsoleNodes(IBacklightDeviceControl instance)
+		{
+			if (instance == null)
+				throw new ArgumentNullException("instance");
+
+			yield break;
+		}
+
+		/// <summary>
+		/// Calls the delegate for each console status item.
+		/// </summary>
+		/// <param name="instance"></param>
+		/// <param name="addRow"></param>
+		public static void BuildConsoleStatus(IBacklightDeviceControl instance, AddStatusRowDelegate addRow)
+		{
+			if (instance == null)
+				throw new ArgumentNullException("instance");
+
+			addRow("BacklightState", instance.BacklightState);
+		}
+
+		/// <summary>
+		/// Gets the child console commands.
+		/// </summary>
+		/// <param name="instance"></param>
+		/// <returns></returns>
+		public static IEnumerable<IConsoleCommand> GetConsoleCommands(IBacklightDeviceControl instance)
+		{
+			if (instance == null)
+				throw new ArgumentNullException("instance");
+
+			yield return new ConsoleCommand("BacklightOn", "Turns on the backlight", () => instance.BacklightOn());
+			yield return new ConsoleCommand("BacklightOff", "Turns off the backlight", () => instance.BacklightOff());
+		}
+	}
+}