+++ ICD.Connect.Panels/Devices/AbstractPanelDevice.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using ICD.Connect.API.Commands;
 using ICD.Connect.API.Nodes;
-using ICD.Connect.Panels.Controls;
 using ICD.Connect.Panels.Controls.Panels;
 using ICD.Connect.Panels.Controls.Sigs;
 using ICD.Connect.Panels.SmartObjectCollections;